+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft/Data/SqlClient/SNI/ConcurrentQueueSemaphore.cs
@@ -15,29 +15,45 @@ namespace Microsoft.Data.SqlClient.SNI
     /// </summary>
     internal sealed partial class ConcurrentQueueSemaphore
     {
-        private static readonly Action<Task, object> s_continuePop = ContinuePop;
-
         private readonly SemaphoreSlim _semaphore;
-        private readonly ConcurrentQueue<TaskCompletionSource<bool>> _queue =
-            new ConcurrentQueue<TaskCompletionSource<bool>>();
+        private readonly ConcurrentQueue<TaskCompletionSource<bool>> _queue;
 
         public ConcurrentQueueSemaphore(int initialCount)
         {
             _semaphore = new SemaphoreSlim(initialCount);
+            _queue = new ConcurrentQueue<TaskCompletionSource<bool>>();
+        }
+
+        public Task WaitAsync(CancellationToken cancellationToken)
+        {
+            // try sync wait with 0 which will not block to see if we need to do an async wait
+            if (_semaphore.Wait(0, cancellationToken))
+            {
+                return Task.CompletedTask;
+            }
+            else
+            {
+                var tcs = new TaskCompletionSource<bool>();
+                _queue.Enqueue(tcs);
+                _semaphore.WaitAsync().ContinueWith(
+                    continuationAction: static (Task task, object state) =>
+                    {
+                        ConcurrentQueue<TaskCompletionSource<bool>> queue = (ConcurrentQueue<TaskCompletionSource<bool>>)state;
+                        if (queue.TryDequeue(out TaskCompletionSource<bool> popped))
+                        {
+                            popped.SetResult(true);
+                        }
+                    },
+                    state: _queue,
+                    cancellationToken: cancellationToken
+                );
+                return tcs.Task;
+            }
         }
 
         public void Release()
         {
             _semaphore.Release();
-        }
-
-        private static void ContinuePop(Task task, object state)
-        {
-            ConcurrentQueue<TaskCompletionSource<bool>> queue = (ConcurrentQueue<TaskCompletionSource<bool>>)state;
-            if (queue.TryDequeue(out TaskCompletionSource<bool> popped))
-            {
-                popped.SetResult(true);
-            }
         }
     }
